+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# [0.15.1](https://github.com/upb-uc4/ui-web/compare/v0.15.0...v0.15.1) (2021-01-07)
+## Feature
+- add frontend validation using backend regex [#755](https://github.com/upb-uc4/ui-web/pull/755)
+- add ability to fetch all information according to gdpr [#776](https://github.com/upb-uc4/ui-web/pull/776)
+- add course admission API [#774](https://github.com/upb-uc4/ui-web/pull/774)
+- add course admission UI [#749](https://github.com/upb-uc4/ui-web/pull/769)
+- add possibility to delete the own account according to gdpr [#765](https://github.com/upb-uc4/ui-web/pull/765)
+
 # [0.15.0](https://github.com/upb-uc4/ui-web/compare/v0.14.0...v0.15.0) (2020-12-18)
 ## Feature
 - add hyperledger versions [#754](https://github.com/upb-uc4/ui-web/pull/754)